requirements.txt
click==8.0.1
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.0
opencv-python==4.5.3.56
pandas==1.3.2
Pillow==8.3.2
PySimpleGUI==4.47.0
python-dateutil==2.8.2
pytz==2021.1
scrypt==0.8.18
six==1.16.0
